requirements.txt
aiogram~=3.20.0.post0
pandas~=2.3.0
python-dotenv~=1.1.0
openpyxl~=3.1.5
SQLAlchemy~=2.0.41
pydantic~=2.11.5